Full editing pass for spring-data-ldap
I edited for spelling, punctuation, grammar, and usage. I also fixed a number of links that were broken because Spring Framework changed it's document structure. I also added the epub cover page and the SVG from which it can be generated. Finally, I removed the introduction.adoc file for the reference section, because its content was unneccessary.